+++ CommunalCalculatorTest/Models/Services/Imp/ColdWaterByMeter.cs
@@ -2,7 +2,7 @@
 {
     internal class ColdWaterByMeter : CommunalServiceByMeter
     {
-        public ColdWaterByMeter(int scopeOfService) : base(scopeOfService) { }
-        public ColdWaterByMeter(int readingBefore, int readingNow) : base(readingBefore, readingNow) { }
+        public ColdWaterByMeter(decimal scopeOfService) : base(scopeOfService) { }
+        public ColdWaterByMeter(decimal readingBefore, decimal readingNow) : base(readingBefore, readingNow) { }
     }
 }
